Removing fingerprint capabilities when enabling or changing PIN
